Add mbed TLS Hashing example (#3)
* Add mbed TLS benchmark example

Add the mbed TLS benchmark example that runs multiple cryptographic
primitives and reports performance. Also, modified the repository's
structure so that each example is in a different subdirectory.

* Fix benchmark code and documentation

Fix the benchmark code to remove the conditional preprocessor
directives for platform.h. This code is not meant to be portable to
anything other than mbed OS, so these are not required.

Also, update documentation according to feedback.

* Add mbed TLS hashing example

Add the mbed TLS example that demonstrates the use of hashing
functions.

* Modify hashing sample to use mbedtls/platform.h