Disable EMAC feature in ublox to fix tls-client
When the macro DEVICE_EMAC is enabled, the program will try to use Wifi
by default. However, the tls-client example only supports Ethernet, so
it fails. To avoid this we use the device_has_remove field in the
mbed_app.json to disable the EMAC feature for UBLOX_EVK_ODIN_W2.