Swap initializations of PSA Crypto and Mbed TLS platform
As PSA Crypto is independent of Mbed TLS, `psa_crypto_init()` must
eventually work standalone, and in particular not require a prior
call to `mbedtls_platform_init()`. At the moment, however, it might
still rely on HW drivers being initialized by `mbedtls_platform_init()`,
and so the latter should be invoked before `psa_crypto_init()`. This
commit changes the order of the calls accordingly.

In general, the relation between of `psa_crypto_init()` and
`mbedtls_platform_init()` still needs to be discussed and documented,
and once that's done, the code might need revision. As it stands,
however, it should work with current implementations of
`psa_crypto_init()` and `mbedtls_platform_init()`.