benchmark: Use benchmark from Mbed TLS
We previously used an older copy of benchmark.c and modified it. This
made it difficult to keep up with improvements and bug fixes from Mbed
TLS. For easier maintenance, use the benchmark.c application as-is (no
modifications) from Mbed TLS. The version imported in this commit is
from Mbed TLS v2.25.0.

Update the expected log output to reflect the new benchmarking output
style.

The benchmark application from Mbed TLS has a dependency on
MBEDTLS_TIMING_C, so add that to our application-specific Mbed TLS
configuration file. This will use the alternate implementation of the
Mbed TLS timing module that Mbed OS provides.

As the benchmark application was designed to run on a PC, its stack
usage is pretty heavy. Configure the main stack to be 32 KiB to
accommodate this.

Fixes #297